.NEW adding a random generate for unique AWS names ( ALB, ECR )
The root-example was extended to generate a unique postfix. This
postfix will be stable during one terraform run. This has the benefit
that a correleation between the compontents of one run can be done.
This postfix is then handed into the modules: nomad, networking to
ensure a unique name of the ALB and ECR configuration in one AWS region.